+++ ntex/src/http/client/pool.rs
@@ -332,7 +332,7 @@ where
                 {
                     if let Some(timeout) = self.disconnect_timeout {
                         if let ConnectionType::H1(io) = conn.io {
-                            crate::rt::spawn(CloseConnection::new(io, timeout))
+                            crate::rt::spawn(CloseConnection::new(io, timeout));
                         }
                     }
                 } else {
@@ -346,7 +346,7 @@ where
                                     if let ConnectionType::H1(io) = io {
                                         crate::rt::spawn(CloseConnection::new(
                                             io, timeout,
-                                        ))
+                                        ));
                                     }
                                 }
                                 continue;
@@ -378,7 +378,7 @@ where
         self.acquired -= 1;
         if let Some(timeout) = self.disconnect_timeout {
             if let ConnectionType::H1(io) = io {
-                crate::rt::spawn(CloseConnection::new(io, timeout))
+                crate::rt::spawn(CloseConnection::new(io, timeout));
             }
         }
         self.check_availibility();
@@ -527,7 +527,7 @@ where
             h2: None,
             rx: Some(rx),
             inner: Some(inner),
-        })
+        });
     }
 }